// issuetracker-client/src/modules/team/types.ts
export type Team = {
    id: string;
    name: string;
    teamMembers: TeamMember[];
};

export type TeamMember = {
    username: string;
    teamId: string;
};

export type User = {
    id: string;
    username: string;
    firstName: string;
    lastName: string;
    dateCreated: Date;
};

export type CreateUserRequest = {
    username: string;
    firstName: string;
    lastName: string;
};